4.21

+++ part4/bloglist/controllers/blogs.js
@@ -46,8 +46,13 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (blogToDelete.user.toString() === decodedToken.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  response.status(400).send({ error: 'only the user can deleted the blog' })
 })
 
 module.exports = blogsRouter
